remove splash page + update nav links
